[zsh] Unify how app/command presence in system is happening. Some minor cleanup.

+++ .dziman-zsh/050_aliases.enabled.zsh
@@ -6,9 +6,18 @@
 
 # General
 alias cls='clear'
-alias ls='gls --classify --human-readable --color=always'
-alias mv='mv -i'
-alias k='k --human'
+
+if command-exists gls; then
+    alias ls='gls --classify --human-readable --color=auto'
+else
+    alias ls='ls -C -p --color=auto'
+fi
+
+if command-exists gmv; then
+    alias mv='gmv --interactive'
+else    
+    alias mv='mv -i'
+fi
 
 # command to track dotfiles by git
 alias config='git --git-dir=$HOME/.dziman-dotfiles/ --work-tree=$HOME'
